requirements.txt
amqp==5.0.6
asgiref==3.4.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
bs4==0.0.1
celery==5.2.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-filter==21.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
idna==3.3
kombu==5.2.2
prompt-toolkit==3.0.23
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==3.5.3
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
